Modularize attachment transfer runtime
Separate cache integrity and download/restore transport from attachment orchestration, organize tests by concern, and preserve the existing command API while satisfying strict linting.
