+++ plugins/attachment-sync/src/runtime.rs
@@ -1,21 +1,18 @@
 #[cfg(test)]
 use std::collections::HashSet;
-use std::io::{Read, Seek, SeekFrom, Write};
 use std::path::{Component, Path, PathBuf};
-use std::time::Duration;
 #[cfg(test)]
-use std::time::SystemTime;
+use std::time::{Duration, SystemTime};
 
 use anlg_e2ee::{
     AttachmentBlobCiphertextMetadata, AttachmentBlobContext, AttachmentBlobMetadata,
     AttachmentBlobPlaintextMetadata, WorkspaceKey,
 };
-use futures_util::StreamExt;
+#[cfg(test)]
 use sha2::{Digest, Sha256};
 use sqlx::FromRow;
-use tauri::{Manager, Runtime};
+use tauri::Runtime;
 use tauri_plugin_settings::SettingsPluginExt;
-use tokio::io::AsyncWriteExt;
 use uuid::{Uuid, Version};
 
 use crate::control::DownloadOperation;
@@ -25,8 +22,24 @@ use crate::models::{
     UploadDescriptor,
 };
 
+mod cache;
 mod delete_guard;
+mod download;
 
+use cache::{
+    CacheFileGuard, cached_shared_attachment_path, cleanup_shared_upload_path,
+    clear_attachment_cache_directory, commit_shared_cache_entry, create_shared_upload_cache_root,
+    ensure_file_size, file_matches_async, file_matches_cancellable_async, private_cache_path,
+    private_cache_root, read_range, read_shared_cache_metadata, shared_cache_id,
+    shared_cache_matches_async, shared_cache_metadata_path, shared_cache_root,
+    shared_preview_cache_root, shared_scope_path, shared_upload_cache_path,
+    shared_upload_cache_root, snapshot_verified_file, valid_cache_id, valid_sha256, validate_range,
+};
+#[cfg(test)]
+use cache::{
+    MAX_RANGE_BYTES, file_matches, hash_identifier, hex_digest, shared_cache_matches,
+    write_shared_cache_metadata,
+};
 #[cfg(test)]
 use delete_guard::{
     DELETE_GUARD_ORPHAN_GRACE, DeleteGuardReconcile, DeleteSourcePreflight,
@@ -35,12 +48,14 @@ use delete_guard::{
     reconcile_staged_delete_guard_once, seal_delete_guard, stage_delete_guard_restore,
 };
 pub use delete_guard::{commit_delete_guard, prepare_delete_guard, reconcile_delete_guards};
+use download::{
+    DownloadObject, download_to_path, persist_staged_attachment, require_configured_supabase_url,
+    stage_attachment_restore, validate_signed_download_url,
+};
 
 const FORMAT_VERSION: i16 = 1;
-const MAX_RANGE_BYTES: u64 = 6 * 1024 * 1024;
 const MAX_PLAINTEXT_BYTES: u64 = anlg_e2ee::ATTACHMENT_BLOB_MAX_PLAINTEXT_BYTES;
 const MAX_CIPHERTEXT_BYTES: u64 = 545_259_520;
-const DOWNLOAD_TIMEOUT: Duration = Duration::from_secs(30 * 60);
 pub(crate) const SHARED_PREVIEW_SCOPE_PREFIX: &str = "preview:";
 
 #[derive(Debug, Clone, FromRow)]
@@ -89,50 +104,6 @@ struct SharedUploadAttachment {
     content_type: String,
     cloud_sync_enabled: i64,
     cloud_object_key: String,
-}
-
-struct CacheFileGuard {
-    path: Option<PathBuf>,
-}
-
-impl CacheFileGuard {
-    fn new(path: PathBuf) -> Self {
-        Self { path: Some(path) }
-    }
-
-    fn disarm(mut self) {
-        self.path = None;
-    }
-}
-
-impl Drop for CacheFileGuard {
-    fn drop(&mut self) {
-        if let Some(path) = self.path.take() {
-            let _ = std::fs::remove_file(path);
-        }
-    }
-}
-
-struct SharedUploadCacheFileGuard {
-    path: Option<PathBuf>,
-}
-
-impl SharedUploadCacheFileGuard {
-    fn new(path: PathBuf) -> Self {
-        Self { path: Some(path) }
-    }
-
-    fn disarm(mut self) {
-        self.path = None;
-    }
-}
-
-impl Drop for SharedUploadCacheFileGuard {
-    fn drop(&mut self) {
-        if let Some(path) = self.path.take() {
-            let _ = cleanup_shared_upload_path(&path);
-        }
-    }
 }
 
 pub async fn describe_upload(
@@ -513,6 +484,7 @@ pub async fn cleanup_shared_upload<R: Runtime>(
         .map_err(|_| Error::CacheUnavailable)?
 }
 
+#[allow(clippy::too_many_arguments)]
 pub async fn download_and_restore<R: Runtime>(
     app: &tauri::AppHandle<R>,
     state: &tauri_plugin_db::ManagedState,
@@ -1318,1444 +1290,5 @@ fn validate_opaque_id(value: &str) -> Result<()> {
     Ok(())
 }
 
-fn valid_sha256(value: &str) -> bool {
-    value.len() == 64
-        && value
-            .bytes()
-            .all(|byte| byte.is_ascii_digit() || matches!(byte, b'a'..=b'f'))
-}
-
-fn valid_cache_id(value: &str) -> bool {
-    Uuid::parse_str(value)
-        .is_ok_and(|uuid| uuid.to_string() == value && uuid.get_version() == Some(Version::Random))
-}
-
-fn validate_range(start: u64, end: u64) -> Result<()> {
-    if end <= start || end - start > MAX_RANGE_BYTES {
-        return Err(Error::InvalidRange);
-    }
-    Ok(())
-}
-
-async fn read_range(path: PathBuf, start: u64, end: u64, expected_size: u64) -> Result<Vec<u8>> {
-    tokio::task::spawn_blocking(move || {
-        let mut file = std::fs::File::open(path)?;
-        if file.metadata()?.len() != expected_size {
-            return Err(Error::CacheUnavailable);
-        }
-        file.seek(SeekFrom::Start(start))?;
-        let length = usize::try_from(end - start).map_err(|_| Error::InvalidRange)?;
-        let mut bytes = vec![0_u8; length];
-        file.read_exact(&mut bytes)?;
-        Ok(bytes)
-    })
-    .await
-    .map_err(|_| Error::CacheUnavailable)?
-}
-
-fn ensure_file_size(path: &Path, expected_size: u64) -> Result<()> {
-    if std::fs::metadata(path)?.len() != expected_size {
-        return Err(Error::LocalAttachmentUnavailable);
-    }
-    Ok(())
-}
-
-fn file_matches(path: &Path, expected_size: u64, expected_sha256: &str) -> Result<bool> {
-    let metadata = match std::fs::metadata(path) {
-        Ok(metadata) => metadata,
-        Err(error) if error.kind() == std::io::ErrorKind::NotFound => return Ok(false),
-        Err(error) => return Err(error.into()),
-    };
-    if !metadata.is_file() || metadata.len() != expected_size {
-        return Ok(false);
-    }
-    Ok(hash_file(path)? == expected_sha256)
-}
-
-async fn file_matches_async(
-    path: PathBuf,
-    expected_size: u64,
-    expected_sha256: String,
-) -> Result<bool> {
-    tokio::task::spawn_blocking(move || file_matches(&path, expected_size, &expected_sha256))
-        .await
-        .map_err(|_| Error::CacheUnavailable)?
-}
-
-async fn file_matches_cancellable_async(
-    path: PathBuf,
-    expected_size: u64,
-    expected_sha256: String,
-    cancellation: tokio_util::sync::CancellationToken,
-) -> Result<bool> {
-    tokio::task::spawn_blocking(move || {
-        file_matches_cancellable(&path, expected_size, &expected_sha256, &cancellation)
-    })
-    .await
-    .map_err(|_| Error::CacheUnavailable)?
-}
-
-fn file_matches_cancellable(
-    path: &Path,
-    expected_size: u64,
-    expected_sha256: &str,
-    cancellation: &tokio_util::sync::CancellationToken,
-) -> Result<bool> {
-    if cancellation.is_cancelled() {
-        return Err(Error::Cancelled);
-    }
-    let metadata = match std::fs::metadata(path) {
-        Ok(metadata) => metadata,
-        Err(error) if error.kind() == std::io::ErrorKind::NotFound => return Ok(false),
-        Err(error) => return Err(error.into()),
-    };
-    if !metadata.is_file() || metadata.len() != expected_size {
-        return Ok(false);
-    }
-    let mut file = std::fs::File::open(path)?;
-    let mut hasher = Sha256::new();
-    let mut buffer = [0_u8; 64 * 1024];
-    loop {
-        if cancellation.is_cancelled() {
-            return Err(Error::Cancelled);
-        }
-        let read = file.read(&mut buffer)?;
-        if read == 0 {
-            break;
-        }
-        hasher.update(&buffer[..read]);
-    }
-    if cancellation.is_cancelled() {
-        return Err(Error::Cancelled);
-    }
-    Ok(hex_digest(hasher.finalize().as_slice()) == expected_sha256)
-}
-
-fn hash_file(path: &Path) -> Result<String> {
-    let mut file = std::fs::File::open(path)?;
-    let mut hasher = Sha256::new();
-    let mut buffer = [0_u8; 64 * 1024];
-    loop {
-        let read = file.read(&mut buffer)?;
-        if read == 0 {
-            break;
-        }
-        hasher.update(&buffer[..read]);
-    }
-    Ok(hex_digest(hasher.finalize().as_slice()))
-}
-
-fn snapshot_verified_file(
-    source_path: &Path,
-    cache_path: &Path,
-    expected_size: u64,
-    expected_sha256: &str,
-    cancellation: &tokio_util::sync::CancellationToken,
-) -> Result<SharedUploadCacheFileGuard> {
-    if cancellation.is_cancelled() {
-        return Err(Error::Cancelled);
-    }
-    let mut source = std::fs::File::open(source_path)?;
-    let source_metadata = source.metadata()?;
-    if !source_metadata.is_file() || source_metadata.len() != expected_size {
-        return Err(Error::LocalAttachmentUnavailable);
-    }
-    let mut options = std::fs::OpenOptions::new();
-    options.create_new(true).write(true);
-    #[cfg(unix)]
-    {
-        use std::os::unix::fs::OpenOptionsExt;
-        options.mode(0o600);
-    }
-    let cache_guard = SharedUploadCacheFileGuard::new(cache_path.to_path_buf());
-    let mut cache = options.open(cache_path)?;
-    let mut hasher = Sha256::new();
-    let mut total = 0_u64;
-    let mut buffer = [0_u8; 64 * 1024];
-    loop {
-        if cancellation.is_cancelled() {
-            return Err(Error::Cancelled);
-        }
-        let read = source.read(&mut buffer)?;
-        if read == 0 {
-            break;
-        }
-        if cancellation.is_cancelled() {
-            return Err(Error::Cancelled);
-        }
-        total = total
-            .checked_add(read as u64)
-            .ok_or(Error::ChecksumMismatch)?;
-        if total > expected_size {
-            return Err(Error::ChecksumMismatch);
-        }
-        hasher.update(&buffer[..read]);
-        cache.write_all(&buffer[..read])?;
-    }
-    if cancellation.is_cancelled() {
-        return Err(Error::Cancelled);
-    }
-    cache.sync_all()?;
-    if total != expected_size || hex_digest(hasher.finalize().as_slice()) != expected_sha256 {
-        return Err(Error::ChecksumMismatch);
-    }
-    Ok(cache_guard)
-}
-
-fn private_cache_root<R: Runtime>(app: &tauri::AppHandle<R>) -> Result<PathBuf> {
-    Ok(app
-        .path()
-        .app_cache_dir()
-        .map_err(|_| Error::CacheUnavailable)?
-        .join("attachment-sync")
-        .join("private"))
-}
-
-fn private_cache_path(root: &Path, cache_id: &str) -> Result<PathBuf> {
-    if !valid_cache_id(cache_id) {
-        return Err(Error::InvalidTransferState);
-    }
-    Ok(root.join(format!("{cache_id}.anb1")))
-}
-
-fn shared_upload_cache_root<R: Runtime>(app: &tauri::AppHandle<R>) -> Result<PathBuf> {
-    Ok(app
-        .path()
-        .app_cache_dir()
-        .map_err(|_| Error::CacheUnavailable)?
-        .join("attachment-sync")
-        .join("shared-upload"))
-}
-
-async fn create_shared_upload_cache_root(path: &Path) -> Result<()> {
-    tokio::fs::create_dir_all(path).await?;
-    #[cfg(unix)]
-    {
-        use std::os::unix::fs::PermissionsExt;
-        tokio::fs::set_permissions(path, std::fs::Permissions::from_mode(0o700)).await?;
-    }
-    Ok(())
-}
-
-fn shared_upload_cache_path(root: &Path, cache_id: &str) -> Result<PathBuf> {
-    if !valid_cache_id(cache_id) {
-        return Err(Error::InvalidTransferState);
-    }
-    Ok(root.join(format!("{cache_id}.bin")))
-}
-
-fn cleanup_shared_upload_path(path: &Path) -> Result<bool> {
-    let mut last_error = None;
-    for delay_ms in [0, 50, 250, 1_000] {
-        if delay_ms > 0 {
-            std::thread::sleep(Duration::from_millis(delay_ms));
-        }
-        match std::fs::remove_file(path) {
-            Ok(()) => return Ok(true),
-            Err(error) if error.kind() == std::io::ErrorKind::NotFound => return Ok(false),
-            Err(error) => last_error = Some(error),
-        }
-    }
-    Err(last_error.ok_or(Error::CacheUnavailable)?.into())
-}
-
-fn shared_scope_path<R: Runtime>(app: &tauri::AppHandle<R>, scope_id: &str) -> Result<PathBuf> {
-    let root = if let Some(view_id) = scope_id.strip_prefix(SHARED_PREVIEW_SCOPE_PREFIX) {
-        if !valid_cache_id(view_id) {
-            return Err(Error::InvalidMetadata);
-        }
-        shared_preview_cache_root(app)?
-    } else {
-        shared_cache_root(app)?
-    };
-    Ok(root.join(hash_identifier(scope_id)))
-}
-
-fn shared_cache_root<R: Runtime>(app: &tauri::AppHandle<R>) -> Result<PathBuf> {
-    Ok(app
-        .path()
-        .app_cache_dir()
-        .map_err(|_| Error::CacheUnavailable)?
-        .join("attachment-sync")
-        .join("shared"))
-}
-
-fn shared_preview_cache_root<R: Runtime>(app: &tauri::AppHandle<R>) -> Result<PathBuf> {
-    Ok(app
-        .path()
-        .app_cache_dir()
-        .map_err(|_| Error::CacheUnavailable)?
-        .join("attachment-sync")
-        .join("shared-preview"))
-}
-
-fn clear_attachment_cache_directory(path: &Path) -> Result<()> {
-    match std::fs::remove_dir_all(path) {
-        Ok(()) => Ok(()),
-        Err(error) if error.kind() == std::io::ErrorKind::NotFound => Ok(()),
-        Err(error) => Err(error.into()),
-    }
-}
-
-fn cached_shared_attachment_path(scope_path: &Path, cache_id: &str) -> PathBuf {
-    scope_path.join(format!("{cache_id}.bin"))
-}
-
-fn shared_cache_metadata_path(scope_path: &Path, cache_id: &str) -> PathBuf {
-    scope_path.join(format!("{cache_id}.meta"))
-}
-
-fn write_shared_cache_metadata(
-    scope_path: &Path,
-    cache_id: &str,
-    size_bytes: u64,
-    sha256: &str,
-) -> Result<()> {
-    let path = shared_cache_metadata_path(scope_path, cache_id);
-    let mut temp = tempfile::NamedTempFile::new_in(scope_path)?;
-    write!(temp, "{size_bytes}\n{sha256}\n")?;
-    temp.flush()?;
-    temp.as_file().sync_all()?;
-    temp.persist(path).map_err(|error| Error::Io(error.error))?;
-    Ok(())
-}
-
-fn commit_shared_cache_entry(
-    scope_path: &Path,
-    cache_id: &str,
-    data_path: &Path,
-    size_bytes: u64,
-    sha256: &str,
-) -> Result<()> {
-    let data_guard = CacheFileGuard::new(data_path.to_path_buf());
-    write_shared_cache_metadata(scope_path, cache_id, size_bytes, sha256)?;
-    data_guard.disarm();
-    Ok(())
-}
-
-fn read_shared_cache_metadata(scope_path: &Path, cache_id: &str) -> Result<Option<(u64, String)>> {
-    let path = shared_cache_metadata_path(scope_path, cache_id);
-    let metadata = match std::fs::metadata(&path) {
-        Ok(metadata) => metadata,
-        Err(error) if error.kind() == std::io::ErrorKind::NotFound => return Ok(None),
-        Err(error) => return Err(error.into()),
-    };
-    if !metadata.is_file() || metadata.len() > 128 {
-        return Ok(None);
-    }
-    let content = match std::fs::read_to_string(path) {
-        Ok(content) => content,
-        Err(error) if error.kind() == std::io::ErrorKind::NotFound => return Ok(None),
-        Err(error) => return Err(error.into()),
-    };
-    let mut lines = content.lines();
-    let size = lines
-        .next()
-        .and_then(|value| value.parse::<u64>().ok())
-        .filter(|value| *value <= MAX_PLAINTEXT_BYTES);
-    let sha256 = lines.next().filter(|value| valid_sha256(value));
-    if lines.next().is_some() || size.is_none() || sha256.is_none() {
-        return Ok(None);
-    }
-    Ok(Some((size.unwrap(), sha256.unwrap().to_string())))
-}
-
-fn shared_cache_matches(
-    scope_path: &Path,
-    cache_id: &str,
-    expected_size: u64,
-    expected_sha256: &str,
-) -> Result<bool> {
-    let Some((recorded_size, recorded_sha256)) = read_shared_cache_metadata(scope_path, cache_id)?
-    else {
-        return Ok(false);
-    };
-    if recorded_size != expected_size || recorded_sha256 != expected_sha256 {
-        return Ok(false);
-    }
-    file_matches(
-        &cached_shared_attachment_path(scope_path, cache_id),
-        expected_size,
-        expected_sha256,
-    )
-}
-
-async fn shared_cache_matches_async(
-    scope_path: PathBuf,
-    cache_id: String,
-    expected_size: u64,
-    expected_sha256: String,
-) -> Result<bool> {
-    tokio::task::spawn_blocking(move || {
-        shared_cache_matches(&scope_path, &cache_id, expected_size, &expected_sha256)
-    })
-    .await
-    .map_err(|_| Error::CacheUnavailable)?
-}
-
-fn hash_identifier(value: &str) -> String {
-    hex_digest(Sha256::digest(value.as_bytes()).as_slice())
-}
-
-fn shared_cache_id(scope_id: &str, attachment_id: &str) -> String {
-    let mut hasher = Sha256::new();
-    hasher.update((scope_id.len() as u64).to_be_bytes());
-    hasher.update(scope_id.as_bytes());
-    hasher.update((attachment_id.len() as u64).to_be_bytes());
-    hasher.update(attachment_id.as_bytes());
-    hex_digest(hasher.finalize().as_slice())
-}
-
-fn hex_digest(bytes: &[u8]) -> String {
-    use std::fmt::Write as _;
-
-    bytes
-        .iter()
-        .fold(String::with_capacity(bytes.len() * 2), |mut value, byte| {
-            write!(&mut value, "{byte:02x}").expect("writing to String cannot fail");
-            value
-        })
-}
-
-enum DownloadObject<'a> {
-    Private(&'a str),
-    Shared(&'a str),
-}
-
-fn require_configured_supabase_url(value: Option<&str>) -> Result<&str> {
-    value
-        .filter(|value| !value.trim().is_empty())
-        .ok_or(Error::InvalidDownloadUrl)
-}
-
-fn validate_signed_download_url(
-    signed_url: &str,
-    supabase_url: &str,
-    object: DownloadObject<'_>,
-) -> Result<url::Url> {
-    let base = url::Url::parse(supabase_url).map_err(|_| Error::InvalidDownloadUrl)?;
-    let signed = url::Url::parse(signed_url).map_err(|_| Error::InvalidDownloadUrl)?;
-    if !base.username().is_empty()
-        || base.password().is_some()
-        || base.query().is_some()
-        || base.fragment().is_some()
-        || !matches!(base.path(), "" | "/")
-        || !signed.username().is_empty()
-        || signed.password().is_some()
-        || signed.fragment().is_some()
-        || base.scheme() != signed.scheme()
-        || base.host_str() != signed.host_str()
-        || base.port_or_known_default() != signed.port_or_known_default()
-        || !secure_or_local(&base)
-    {
-        return Err(Error::InvalidDownloadUrl);
-    }
-    let mut query = signed.query_pairs();
-    if !query
-        .next()
-        .is_some_and(|(name, value)| name == "token" && !value.is_empty())
-        || query.next().is_some()
-    {
-        return Err(Error::InvalidDownloadUrl);
-    }
-    let valid_path = match object {
-        DownloadObject::Private(object_key) => {
-            signed.path() == format!("/storage/v1/object/sign/attachment-backups/{object_key}")
-        }
-        DownloadObject::Shared(attachment_id) => {
-            valid_shared_attachment_download_path(signed.path(), attachment_id)
-        }
-    };
-    if !valid_path {
-        return Err(Error::InvalidDownloadUrl);
-    }
-    Ok(signed)
-}
-
-fn valid_shared_attachment_download_path(path: &str, attachment_id: &str) -> bool {
-    let Some(object_key) = path.strip_prefix("/storage/v1/object/sign/shared-note-attachments/")
-    else {
-        return false;
-    };
-    let mut parts = object_key.split('/');
-    let owner = parts.next();
-    let share = parts.next();
-    let filename = parts.next();
-    parts.next().is_none()
-        && owner.is_some_and(valid_canonical_uuid)
-        && share.is_some_and(valid_canonical_uuid)
-        && filename.is_some_and(|filename| filename == format!("{attachment_id}.sna1"))
-}
-
-fn valid_canonical_uuid(value: &str) -> bool {
-    Uuid::parse_str(value).is_ok_and(|uuid| uuid.to_string() == value)
-}
-
-fn secure_or_local(url: &url::Url) -> bool {
-    if url.scheme() == "https" {
-        return true;
-    }
-    if url.scheme() != "http" {
-        return false;
-    }
-    match url.host() {
-        Some(url::Host::Domain("localhost")) => true,
-        Some(url::Host::Ipv4(address)) => address.is_loopback(),
-        Some(url::Host::Ipv6(address)) => address.is_loopback(),
-        _ => false,
-    }
-}
-
-async fn download_to_path(
-    url: url::Url,
-    path: &Path,
-    expected_size: u64,
-    expected_sha256: &str,
-    create_new: bool,
-    cancellation: &tokio_util::sync::CancellationToken,
-) -> Result<()> {
-    let client = reqwest::Client::builder()
-        .timeout(DOWNLOAD_TIMEOUT)
-        .redirect(reqwest::redirect::Policy::none())
-        .build()
-        .map_err(Error::Download)?;
-    let response = tokio::select! {
-        _ = cancellation.cancelled() => return Err(Error::Cancelled),
-        response = client
-            .get(url)
-            .header(reqwest::header::ACCEPT_ENCODING, "identity")
-            .send() => response.map_err(Error::Download)?,
-    };
-    if !response.status().is_success()
-        || response
-            .content_length()
-            .is_some_and(|size| size != expected_size)
-        || response
-            .headers()
-            .get(reqwest::header::CONTENT_ENCODING)
-            .is_some_and(|value| value != "identity")
-    {
-        return Err(Error::IncompleteDownload);
-    }
-
-    let mut options = tokio::fs::OpenOptions::new();
-    options.write(true).truncate(!create_new);
-    if create_new {
-        options.create_new(true);
-    }
-    let mut file = options.open(path).await?;
-    let mut stream = response.bytes_stream();
-    let mut size = 0_u64;
-    let mut hasher = Sha256::new();
-    loop {
-        let next = tokio::select! {
-            _ = cancellation.cancelled() => return Err(Error::Cancelled),
-            next = stream.next() => next,
-        };
-        let Some(chunk) = next else {
-            break;
-        };
-        let chunk = chunk.map_err(Error::Download)?;
-        size = size
-            .checked_add(chunk.len() as u64)
-            .ok_or(Error::IncompleteDownload)?;
-        if size > expected_size {
-            return Err(Error::IncompleteDownload);
-        }
-        hasher.update(&chunk);
-        file.write_all(&chunk).await?;
-    }
-    tokio::select! {
-        _ = cancellation.cancelled() => return Err(Error::Cancelled),
-        result = file.sync_all() => result?,
-    }
-    if size != expected_size || hex_digest(hasher.finalize().as_slice()) != expected_sha256 {
-        return Err(Error::ChecksumMismatch);
-    }
-    Ok(())
-}
-
-fn stage_attachment_restore(
-    key: &WorkspaceKey,
-    context: &AttachmentBlobContext,
-    cache_path: &Path,
-    destination_parent: &Path,
-    expected: &AttachmentBlobMetadata,
-) -> Result<tempfile::NamedTempFile> {
-    std::fs::create_dir_all(destination_parent)?;
-    let mut source = std::fs::File::open(cache_path)?;
-    let mut temp = tempfile::NamedTempFile::new_in(destination_parent)?;
-    key.open_attachment_blob(context, &mut source, &mut temp, expected)?;
-    temp.flush()?;
-    temp.as_file().sync_all()?;
-    Ok(temp)
-}
-
-fn persist_staged_attachment(staged: tempfile::NamedTempFile, destination: &Path) -> Result<()> {
-    let parent = destination
-        .parent()
-        .ok_or(Error::LocalAttachmentUnavailable)?;
-    staged
-        .persist(destination)
-        .map(|_| ())
-        .map_err(|error| Error::Io(error.error))?;
-    sync_destination_directory(parent)
-}
-
-#[cfg(unix)]
-fn sync_destination_directory(path: &Path) -> Result<()> {
-    std::fs::File::open(path)?.sync_all()?;
-    Ok(())
-}
-
-#[cfg(not(unix))]
-fn sync_destination_directory(_path: &Path) -> Result<()> {
-    Ok(())
-}
-
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    fn transfer_record(cloud_sync_enabled: i64) -> TransferAttachment {
-        TransferAttachment {
-            job_id: "job-1".to_string(),
-            attachment_id: "attachment-1".to_string(),
-            session_id: "session-1".to_string(),
-            workspace_id: "workspace-1".to_string(),
-            expected_sha256: "a".repeat(64),
-            expected_size_bytes: 1,
-            ciphertext_sha256: String::new(),
-            ciphertext_size_bytes: 0,
-            remote_object_id: String::new(),
-            object_key: String::new(),
-            cache_id: String::new(),
-            phase: "preparing".to_string(),
-            relative_path: "attachments/file.bin".to_string(),
-            source_type: "note_upload".to_string(),
-            attachment_sha256: "a".repeat(64),
-            attachment_size_bytes: 1,
-            attachment_cloud_object_key: String::new(),
-            cloud_sync_enabled,
-        }
-    }
-
-    fn delete_source_record(cloud_sync_enabled: i64) -> DeleteSourcePreflight {
-        DeleteSourcePreflight {
-            attachment_id: "attachment-1".to_string(),
-            session_id: "session-1".to_string(),
-            workspace_id: "workspace-1".to_string(),
-            expected_sha256: "a".repeat(64),
-            expected_size_bytes: 1,
-            object_key: "owner/object.anb1".to_string(),
-            cache_id: String::new(),
-            ciphertext_sha256: String::new(),
-            ciphertext_size_bytes: 0,
-            current_attachment_id: Some("attachment-1".to_string()),
-            relative_path: Some("attachments/file.bin".to_string()),
-            source_type: Some("note_upload".to_string()),
-            attachment_sha256: Some("a".repeat(64)),
-            attachment_size_bytes: Some(1),
-            attachment_cloud_object_key: Some("owner/object.anb1".to_string()),
-            cloud_sync_enabled: Some(cloud_sync_enabled),
-            deleted_at: None,
-            local_availability: "present".to_string(),
-        }
-    }
-
-    #[test]
-    fn disabled_cloud_sync_is_restore_only() {
-        let disabled = transfer_record(0);
-        assert!(validate_transfer_version(&disabled).is_ok());
-        assert!(validate_upload_transfer_version(&disabled).is_err());
-
-        let enabled = transfer_record(1);
-        assert!(validate_transfer_version(&enabled).is_ok());
-        assert!(validate_upload_transfer_version(&enabled).is_ok());
-    }
-
-    #[test]
-    fn only_an_exact_current_delete_source_requires_hashing() {
-        assert!(
-            delete_source_attachment(&delete_source_record(0))
-                .unwrap()
-                .is_some()
-        );
-        assert!(
-            delete_source_attachment(&delete_source_record(1))
-                .unwrap()
-                .is_some()
-        );
-
-        let mut missing = delete_source_record(0);
-        missing.current_attachment_id = None;
-        assert!(delete_source_attachment(&missing).unwrap().is_none());
-
-        let mut deleted = delete_source_record(0);
-        deleted.deleted_at = Some("2026-07-18T00:00:00.000Z".to_string());
-        deleted.relative_path = None;
-        assert!(delete_source_attachment(&deleted).unwrap().is_none());
-
-        let mut replaced = delete_source_record(0);
-        replaced.attachment_sha256 = Some("b".repeat(64));
-        assert!(delete_source_attachment(&replaced).unwrap().is_none());
-
-        let mut other_object = delete_source_record(0);
-        other_object.attachment_cloud_object_key = Some("owner/other.anb1".to_string());
-        assert!(delete_source_attachment(&other_object).unwrap().is_none());
-    }
-
-    #[test]
-    fn exact_remote_dependency_blocks_delete_commit() {
-        let mut enabled = delete_source_record(1);
-        assert!(exact_delete_dependency(&enabled).unwrap());
-
-        enabled.cloud_sync_enabled = Some(0);
-        enabled.local_availability = "absent".to_string();
-        assert!(exact_delete_dependency(&enabled).unwrap());
-
-        enabled.local_availability = "present".to_string();
-        assert!(!exact_delete_dependency(&enabled).unwrap());
-
-        enabled.attachment_cloud_object_key = Some("owner/newer.anb1".to_string());
-        assert!(!exact_delete_dependency(&enabled).unwrap());
-    }
-
-    #[test]
-    fn delete_backup_refs_are_stable_for_persisted_job_metadata() {
-        let key = anlg_e2ee::RecoveryKey::generate()
-            .unwrap()
-            .workspace_key("workspace-1")
-            .unwrap();
-        let expected = plaintext_metadata(&"a".repeat(64), 42).unwrap();
-        let first = attachment_backup_refs(&key, "workspace-1", "attachment-1", &expected).unwrap();
-        let retry = attachment_backup_refs(&key, "workspace-1", "attachment-1", &expected).unwrap();
-        assert_eq!(first, retry);
-
-        let newer = plaintext_metadata(&"b".repeat(64), 42).unwrap();
-        let newer = attachment_backup_refs(&key, "workspace-1", "attachment-1", &newer).unwrap();
-        assert_eq!(first.0, newer.0);
-        assert_ne!(first.1, newer.1);
-    }
-
-    #[test]
-    fn delete_source_hash_detects_missing_and_changed_local_bytes() {
-        let directory = tempfile::tempdir().unwrap();
-        let path = directory.path().join("attachment.bin");
-        let bytes = b"preserved attachment";
-        let sha256 = hex_digest(Sha256::digest(bytes).as_slice());
-
-        assert!(!file_matches(&path, bytes.len() as u64, &sha256).unwrap());
-        std::fs::write(&path, bytes).unwrap();
-        assert!(file_matches(&path, bytes.len() as u64, &sha256).unwrap());
-        std::fs::write(&path, b"different attachment").unwrap();
-        assert!(!file_matches(&path, bytes.len() as u64, &sha256).unwrap());
-    }
-
-    #[test]
-    fn shared_upload_snapshot_stays_stable_when_the_source_changes() {
-        let directory = tempfile::tempdir().unwrap();
-        let source = directory.path().join("attachment.bin");
-        let snapshot = directory.path().join("snapshot.bin");
-        let bytes = b"stable shared attachment";
-        let sha256 = hex_digest(Sha256::digest(bytes).as_slice());
-        std::fs::write(&source, bytes).unwrap();
-
-        snapshot_verified_file(
-            &source,
-            &snapshot,
-            bytes.len() as u64,
-            &sha256,
-            &tokio_util::sync::CancellationToken::new(),
-        )
-        .unwrap()
-        .disarm();
-        std::fs::write(&source, vec![b'x'; bytes.len()]).unwrap();
-
-        assert_eq!(std::fs::read(&snapshot).unwrap(), bytes);
-        assert!(file_matches(&snapshot, bytes.len() as u64, &sha256).unwrap());
-        assert!(!file_matches(&source, bytes.len() as u64, &sha256).unwrap());
-        #[cfg(unix)]
-        {
-            use std::os::unix::fs::PermissionsExt;
-            assert_eq!(
-                std::fs::metadata(&snapshot).unwrap().permissions().mode() & 0o777,
-                0o600
-            );
-        }
-    }
-
-    #[test]
-    fn shared_upload_snapshot_rejects_and_removes_unregistered_bytes() {
-        let directory = tempfile::tempdir().unwrap();
-        let source = directory.path().join("attachment.bin");
-        let snapshot = directory.path().join("snapshot.bin");
-        let expected = b"expected shared attachment";
-        let sha256 = hex_digest(Sha256::digest(expected).as_slice());
-        std::fs::write(&source, vec![b'x'; expected.len()]).unwrap();
-
-        assert!(matches!(
-            snapshot_verified_file(
-                &source,
-                &snapshot,
-                expected.len() as u64,
-                &sha256,
-                &tokio_util::sync::CancellationToken::new(),
-            ),
-            Err(Error::ChecksumMismatch)
-        ));
-        assert!(!snapshot.exists());
-    }
-
-    #[test]
-    fn cancelled_shared_upload_snapshot_leaves_no_plaintext_cache() {
-        let directory = tempfile::tempdir().unwrap();
-        let source = directory.path().join("attachment.bin");
-        let snapshot = directory.path().join("snapshot.bin");
-        let bytes = b"cancelled shared attachment";
-        let sha256 = hex_digest(Sha256::digest(bytes).as_slice());
-        let cancellation = tokio_util::sync::CancellationToken::new();
-        cancellation.cancel();
-        std::fs::write(&source, bytes).unwrap();
-
-        assert!(matches!(
-            snapshot_verified_file(
-                &source,
-                &snapshot,
-                bytes.len() as u64,
-                &sha256,
-                &cancellation,
-            ),
-            Err(Error::Cancelled)
-        ));
-        assert!(!snapshot.exists());
-    }
-
-    #[test]
-    fn shared_upload_cleanup_is_idempotent() {
-        let directory = tempfile::tempdir().unwrap();
-        let snapshot = directory.path().join("snapshot.bin");
-        std::fs::write(&snapshot, b"shared attachment").unwrap();
-
-        assert!(cleanup_shared_upload_path(&snapshot).unwrap());
-        assert!(!cleanup_shared_upload_path(&snapshot).unwrap());
-    }
-
-    #[tokio::test]
-    async fn shared_upload_cache_directory_is_owner_only() {
-        let directory = tempfile::tempdir().unwrap();
-        let root = directory.path().join("shared-upload");
-
-        create_shared_upload_cache_root(&root).await.unwrap();
-        assert!(root.is_dir());
-        #[cfg(unix)]
-        {
-            use std::os::unix::fs::PermissionsExt;
-            assert_eq!(
-                std::fs::metadata(root).unwrap().permissions().mode() & 0o777,
-                0o700
-            );
-        }
-    }
-
-    #[test]
-    fn validates_upload_ranges() {
-        assert!(validate_range(0, 1).is_ok());
-        assert!(validate_range(0, MAX_RANGE_BYTES).is_ok());
-        assert!(validate_range(0, 0).is_err());
-        assert!(validate_range(2, 1).is_err());
-        assert!(validate_range(0, MAX_RANGE_BYTES + 1).is_err());
-    }
-
-    #[test]
-    fn only_accepts_expected_attachment_paths() {
-        assert!(validate_attachment_relative_path("note_upload", "attachments/image.png").is_ok());
-        assert!(validate_attachment_relative_path("session_audio", "audio.wav").is_ok());
-        assert!(
-            validate_attachment_relative_path("session_audio", "attachments/audio.wav").is_err()
-        );
-        assert!(validate_attachment_relative_path("note_upload", "../image.png").is_err());
-        assert!(validate_attachment_relative_path("note_upload", "attachments/a/b.png").is_err());
-    }
-
-    #[test]
-    fn signed_urls_are_origin_and_path_bound() {
-        let object_key =
-            "00000000-0000-4000-8000-000000000001/00000000-0000-4000-8000-000000000002.anb1";
-        let valid = format!(
-            "https://project.supabase.co/storage/v1/object/sign/attachment-backups/{object_key}?token=secret"
-        );
-        assert!(
-            validate_signed_download_url(
-                &valid,
-                "https://project.supabase.co",
-                DownloadObject::Private(object_key),
-            )
-            .is_ok()
-        );
-        assert!(
-            validate_signed_download_url(
-                &valid,
-                "https://other.supabase.co",
-                DownloadObject::Private(object_key),
-            )
-            .is_err()
-        );
-        assert!(
-            validate_signed_download_url(
-                "http://project.supabase.co/storage/v1/object/sign/bucket/item?token=secret",
-                "http://project.supabase.co",
-                DownloadObject::Private(object_key),
-            )
-            .is_err()
-        );
-        assert!(
-            validate_signed_download_url(
-                &format!(
-                    "http://127.0.0.1:54321/storage/v1/object/sign/attachment-backups/{object_key}?token=secret"
-                ),
-                "http://127.0.0.1:54321",
-                DownloadObject::Private(object_key),
-            )
-            .is_ok()
-        );
-
-        let attachment_id = "00000000-0000-4000-8000-000000000003";
-        let shared = format!(
-            "https://project.supabase.co/storage/v1/object/sign/shared-note-attachments/00000000-0000-4000-8000-000000000001/00000000-0000-4000-8000-000000000002/{attachment_id}.sna1?token=secret"
-        );
-        assert!(
-            validate_signed_download_url(
-                &shared,
-                "https://project.supabase.co",
-                DownloadObject::Shared(attachment_id),
-            )
-            .is_ok()
-        );
-        assert!(
-            validate_signed_download_url(
-                &shared.replace("shared-note-attachments", "other-bucket"),
-                "https://project.supabase.co",
-                DownloadObject::Shared(attachment_id),
-            )
-            .is_err()
-        );
-        assert!(
-            validate_signed_download_url(
-                &shared.replace(attachment_id, "00000000-0000-4000-8000-000000000004"),
-                "https://project.supabase.co",
-                DownloadObject::Shared(attachment_id),
-            )
-            .is_err()
-        );
-    }
-
-    #[test]
-    fn missing_or_empty_configured_origins_fail_closed() {
-        assert!(require_configured_supabase_url(None).is_err());
-        assert!(require_configured_supabase_url(Some("")).is_err());
-        assert!(require_configured_supabase_url(Some("   ")).is_err());
-        assert_eq!(
-            require_configured_supabase_url(Some("https://project.supabase.co")).unwrap(),
-            "https://project.supabase.co"
-        );
-    }
-
-    #[test]
-    fn scoped_cache_names_do_not_expose_ids() {
-        let first = hash_identifier("share/../../secret");
-        let second = hash_identifier("share/../../secret");
-        assert_eq!(first, second);
-        assert_eq!(first.len(), 64);
-        assert!(first.bytes().all(|byte| byte.is_ascii_hexdigit()));
-        assert!(!first.contains("secret"));
-        assert_ne!(
-            shared_cache_id("viewer-a", "attachment-a"),
-            shared_cache_id("viewer-b", "attachment-a")
-        );
-    }
-
-    #[test]
-    fn restores_encrypted_attachment_atomically() {
-        let directory = tempfile::tempdir().unwrap();
-        let source_path = directory.path().join("source.bin");
-        let cache_path = directory.path().join("cache.anb1");
-        let destination_path = directory.path().join("destination.bin");
-        let plaintext = b"private attachment bytes";
-        std::fs::write(&source_path, plaintext).unwrap();
-        std::fs::write(&destination_path, b"old bytes").unwrap();
-
-        let key = anlg_e2ee::RecoveryKey::generate()
-            .unwrap()
-            .workspace_key("workspace-a")
-            .unwrap();
-        let context =
-            AttachmentBlobContext::new("workspace-a", "attachment-a", Uuid::new_v4().to_string())
-                .unwrap();
-        let expected_plaintext = AttachmentBlobPlaintextMetadata::from_hex(
-            plaintext.len() as u64,
-            &hex_digest(Sha256::digest(plaintext).as_slice()),
-        )
-        .unwrap();
-        let metadata = {
-            let mut source = std::fs::File::open(source_path).unwrap();
-            let mut cache = std::fs::File::create(&cache_path).unwrap();
-            key.seal_attachment_blob(&context, &mut source, &mut cache, &expected_plaintext)
-                .unwrap()
-        };
-
-        let staged = stage_attachment_restore(
-            &key,
-            &context,
-            &cache_path,
-            destination_path.parent().unwrap(),
-            &metadata,
-        )
-        .unwrap();
-        assert_eq!(std::fs::read(&destination_path).unwrap(), b"old bytes");
-        persist_staged_attachment(staged, &destination_path).unwrap();
-        assert_eq!(std::fs::read(destination_path).unwrap(), plaintext);
-    }
-
-    #[test]
-    fn delete_guard_restores_canonical_bytes_and_preserves_a_conflict() {
-        let directory = tempfile::tempdir().unwrap();
-        let source_path = directory.path().join("source.bin");
-        let guard_path = directory.path().join(format!("{}.anb1", Uuid::new_v4()));
-        let destination_path = directory.path().join("destination.bin");
-        let canonical = b"canonical private attachment";
-        let local_edit = b"different local attachment";
-        std::fs::write(&source_path, canonical).unwrap();
-
-        let key = anlg_e2ee::RecoveryKey::generate()
-            .unwrap()
-            .workspace_key("workspace-a")
-            .unwrap();
-        let context =
-            AttachmentBlobContext::new("workspace-a", "attachment-a", Uuid::new_v4().to_string())
-                .unwrap();
-        let expected = AttachmentBlobPlaintextMetadata::from_hex(
-            canonical.len() as u64,
-            &hex_digest(Sha256::digest(canonical).as_slice()),
-        )
-        .unwrap();
-        let (metadata, guard) = seal_delete_guard(
-            &key,
-            &context,
-            &source_path,
-            &guard_path,
-            &expected,
-            &tokio_util::sync::CancellationToken::new(),
-        )
-        .unwrap();
-        guard.disarm();
-
-        assert_ne!(std::fs::read(&guard_path).unwrap(), canonical);
-        #[cfg(unix)]
-        {
-            use std::os::unix::fs::PermissionsExt;
-            assert_eq!(
-                std::fs::metadata(&guard_path).unwrap().permissions().mode() & 0o777,
-                0o600
-            );
-        }
-
-        std::fs::write(&source_path, b"source changed after delete began").unwrap();
-        std::fs::write(&destination_path, local_edit).unwrap();
-        let staged = stage_delete_guard_restore(
-            &key,
-            &context,
-            &guard_path,
-            directory.path(),
-            &metadata,
-            &tokio_util::sync::CancellationToken::new(),
-        )
-        .unwrap();
-        let conflicts = reconcile_staged_delete_guard(
-            staged,
-            &destination_path,
-            canonical.len() as u64,
-            &expected.sha256_hex(),
-        )
-        .unwrap();
-
-        assert_eq!(std::fs::read(&destination_path).unwrap(), canonical);
-        assert_eq!(conflicts.len(), 1);
-        assert_eq!(std::fs::read(&conflicts[0]).unwrap(), local_edit);
-
-        let staged = stage_delete_guard_restore(
-            &key,
-            &context,
-            &guard_path,
-            directory.path(),
-            &metadata,
-            &tokio_util::sync::CancellationToken::new(),
-        )
-        .unwrap();
-        assert!(
-            reconcile_staged_delete_guard(
-                staged,
-                &destination_path,
-                canonical.len() as u64,
-                &expected.sha256_hex(),
-            )
-            .unwrap()
-            .is_empty()
-        );
-    }
-
-    #[test]
-    fn delete_guard_restores_a_missing_destination() {
-        let directory = tempfile::tempdir().unwrap();
-        let source_path = directory.path().join("source.bin");
-        let guard_path = directory.path().join(format!("{}.anb1", Uuid::new_v4()));
-        let destination_path = directory.path().join("missing.bin");
-        let canonical = b"attachment recovered after remote delete";
-        std::fs::write(&source_path, canonical).unwrap();
-        let key = anlg_e2ee::RecoveryKey::generate()
-            .unwrap()
-            .workspace_key("workspace-a")
-            .unwrap();
-        let context =
-            AttachmentBlobContext::new("workspace-a", "attachment-a", Uuid::new_v4().to_string())
-                .unwrap();
-        let expected = AttachmentBlobPlaintextMetadata::from_hex(
-            canonical.len() as u64,
-            &hex_digest(Sha256::digest(canonical).as_slice()),
-        )
-        .unwrap();
-        let (metadata, guard) = seal_delete_guard(
-            &key,
-            &context,
-            &source_path,
-            &guard_path,
-            &expected,
-            &tokio_util::sync::CancellationToken::new(),
-        )
-        .unwrap();
-        guard.disarm();
-        let staged = stage_delete_guard_restore(
-            &key,
-            &context,
-            &guard_path,
-            directory.path(),
-            &metadata,
-            &tokio_util::sync::CancellationToken::new(),
-        )
-        .unwrap();
-
-        assert!(
-            reconcile_staged_delete_guard(
-                staged,
-                &destination_path,
-                canonical.len() as u64,
-                &expected.sha256_hex(),
-            )
-            .unwrap()
-            .is_empty()
-        );
-        assert_eq!(std::fs::read(destination_path).unwrap(), canonical);
-    }
-
-    #[test]
-    fn delete_guard_retry_keeps_the_plaintext_stage() {
-        let directory = tempfile::tempdir().unwrap();
-        let blocked_parent = directory.path().join("blocked-parent");
-        let destination = blocked_parent.join("attachment.bin");
-        let canonical = b"attachment retained across filesystem retry";
-        let expected_sha256 = hex_digest(Sha256::digest(canonical).as_slice());
-        let staged = tempfile::NamedTempFile::new_in(directory.path()).unwrap();
-        std::fs::write(staged.path(), canonical).unwrap();
-        std::fs::write(&blocked_parent, b"not a directory").unwrap();
-
-        let (staged, conflicts) = match reconcile_staged_delete_guard_once(
-            staged,
-            &destination,
-            canonical.len() as u64,
-            &expected_sha256,
-            Vec::new(),
-        )
-        .unwrap()
-        {
-            DeleteGuardReconcile::Retry {
-                staged,
-                conflicts,
-                error: Error::Io(_),
-            } => (staged, conflicts),
-            _ => panic!("blocked parent should produce a retryable filesystem error"),
-        };
-        assert!(staged.path().exists());
-
-        std::fs::remove_file(blocked_parent).unwrap();
-        let conflicts = match reconcile_staged_delete_guard_once(
-            staged,
-            &destination,
-            canonical.len() as u64,
-            &expected_sha256,
-            conflicts,
-        )
-        .unwrap()
-        {
-            DeleteGuardReconcile::Ready(conflicts) => conflicts,
-            DeleteGuardReconcile::Retry { .. } => panic!("retry should restore the attachment"),
-        };
-
-        assert!(conflicts.is_empty());
-        assert_eq!(std::fs::read(destination).unwrap(), canonical);
-    }
-
-    #[test]
-    fn cancelled_delete_guard_seal_removes_partial_ciphertext() {
-        let directory = tempfile::tempdir().unwrap();
-        let source_path = directory.path().join("source.bin");
-        let guard_path = directory.path().join(format!("{}.anb1", Uuid::new_v4()));
-        let canonical = b"attachment whose delete was cancelled";
-        std::fs::write(&source_path, canonical).unwrap();
-        let key = anlg_e2ee::RecoveryKey::generate()
-            .unwrap()
-            .workspace_key("workspace-a")
-            .unwrap();
-        let context =
-            AttachmentBlobContext::new("workspace-a", "attachment-a", Uuid::new_v4().to_string())
-                .unwrap();
-        let expected = AttachmentBlobPlaintextMetadata::from_hex(
-            canonical.len() as u64,
-            &hex_digest(Sha256::digest(canonical).as_slice()),
-        )
-        .unwrap();
-        let cancellation = tokio_util::sync::CancellationToken::new();
-        cancellation.cancel();
-
-        assert!(matches!(
-            seal_delete_guard(
-                &key,
-                &context,
-                &source_path,
-                &guard_path,
-                &expected,
-                &cancellation,
-            ),
-            Err(Error::Cancelled)
-        ));
-        assert!(!guard_path.exists());
-    }
-
-    #[test]
-    fn cancelled_delete_guard_restore_removes_plaintext_stage() {
-        let directory = tempfile::tempdir().unwrap();
-        let source_path = directory.path().join("source.bin");
-        let guard_path = directory.path().join(format!("{}.anb1", Uuid::new_v4()));
-        let canonical = b"attachment whose restore was cancelled";
-        std::fs::write(&source_path, canonical).unwrap();
-        let key = anlg_e2ee::RecoveryKey::generate()
-            .unwrap()
-            .workspace_key("workspace-a")
-            .unwrap();
-        let context =
-            AttachmentBlobContext::new("workspace-a", "attachment-a", Uuid::new_v4().to_string())
-                .unwrap();
-        let expected = AttachmentBlobPlaintextMetadata::from_hex(
-            canonical.len() as u64,
-            &hex_digest(Sha256::digest(canonical).as_slice()),
-        )
-        .unwrap();
-        let (metadata, guard) = seal_delete_guard(
-            &key,
-            &context,
-            &source_path,
-            &guard_path,
-            &expected,
-            &tokio_util::sync::CancellationToken::new(),
-        )
-        .unwrap();
-        guard.disarm();
-        let cancellation = tokio_util::sync::CancellationToken::new();
-        cancellation.cancel();
-
-        assert!(matches!(
-            stage_delete_guard_restore(
-                &key,
-                &context,
-                &guard_path,
-                directory.path(),
-                &metadata,
-                &cancellation,
-            ),
-            Err(Error::Cancelled)
-        ));
-        assert_eq!(std::fs::read_dir(directory.path()).unwrap().count(), 2);
-    }
-
-    #[test]
-    fn recoverable_delete_guard_drift_uses_a_retryable_message() {
-        let message = Error::DeleteGuardChanged.to_string().to_ascii_lowercase();
-        for permanent_marker in ["invalid", "mismatch", "path", "source"] {
-            assert!(!message.contains(permanent_marker));
-        }
-    }
-
-    #[tokio::test]
-    async fn delete_guard_directory_is_owner_only() {
-        let directory = tempfile::tempdir().unwrap();
-        let root = directory.path().join("delete-guards");
-
-        create_delete_guard_root(&root).await.unwrap();
-        assert!(root.is_dir());
-        #[cfg(unix)]
-        {
-            use std::os::unix::fs::PermissionsExt;
-            assert_eq!(
-                std::fs::metadata(root).unwrap().permissions().mode() & 0o777,
-                0o700
-            );
-        }
-    }
-
-    #[test]
-    fn delete_guard_reconciliation_retains_only_live_references() {
-        let directory = tempfile::tempdir().unwrap();
-        let live = Uuid::new_v4().to_string();
-        let orphan = Uuid::new_v4().to_string();
-        std::fs::write(directory.path().join(format!("{live}.anb1")), b"live").unwrap();
-        std::fs::write(directory.path().join(format!("{orphan}.anb1")), b"orphan").unwrap();
-        std::fs::write(directory.path().join("malformed.anb1"), b"malformed").unwrap();
-
-        assert_eq!(
-            reconcile_delete_guard_files(
-                directory.path(),
-                &HashSet::from([live.clone()]),
-                SystemTime::now() + Duration::from_secs(1),
-            )
-            .unwrap(),
-            2
-        );
-        assert!(directory.path().join(format!("{live}.anb1")).is_file());
-        assert!(!directory.path().join(format!("{orphan}.anb1")).exists());
-        assert!(!directory.path().join("malformed.anb1").exists());
-    }
-
-    #[test]
-    fn delete_guard_reconciliation_does_not_race_a_new_guard() {
-        let directory = tempfile::tempdir().unwrap();
-        let guard_id = Uuid::new_v4().to_string();
-        let guard_path = directory.path().join(format!("{guard_id}.anb1"));
-        std::fs::write(&guard_path, b"new unlinked guard").unwrap();
-
-        assert_eq!(
-            reconcile_delete_guard_files(
-                directory.path(),
-                &HashSet::new(),
-                SystemTime::now() - DELETE_GUARD_ORPHAN_GRACE,
-            )
-            .unwrap(),
-            0
-        );
-        assert!(guard_path.is_file());
-    }
-
-    #[test]
-    fn shared_cache_requires_matching_sidecar_and_bytes() {
-        let directory = tempfile::tempdir().unwrap();
-        let cache_id = shared_cache_id("viewer-a", "attachment-a");
-        let bytes = b"shared attachment";
-        let sha256 = hex_digest(Sha256::digest(bytes).as_slice());
-        std::fs::write(
-            cached_shared_attachment_path(directory.path(), &cache_id),
-            bytes,
-        )
-        .unwrap();
-        write_shared_cache_metadata(directory.path(), &cache_id, bytes.len() as u64, &sha256)
-            .unwrap();
-        assert!(
-            shared_cache_matches(directory.path(), &cache_id, bytes.len() as u64, &sha256).unwrap()
-        );
-
-        std::fs::write(
-            cached_shared_attachment_path(directory.path(), &cache_id),
-            b"tampered attachment",
-        )
-        .unwrap();
-        assert!(
-            !shared_cache_matches(directory.path(), &cache_id, bytes.len() as u64, &sha256)
-                .unwrap()
-        );
-    }
-    #[test]
-    fn shared_cache_removes_plaintext_when_sidecar_commit_fails() {
-        let directory = tempfile::tempdir().unwrap();
-        let cache_id = shared_cache_id("viewer-a", "attachment-a");
-        let data_path = cached_shared_attachment_path(directory.path(), &cache_id);
-        std::fs::write(&data_path, b"shared attachment").unwrap();
-        std::fs::create_dir(shared_cache_metadata_path(directory.path(), &cache_id)).unwrap();
-
-        assert!(
-            commit_shared_cache_entry(
-                directory.path(),
-                &cache_id,
-                &data_path,
-                17,
-                &"a".repeat(64),
-            )
-            .is_err()
-        );
-        assert!(!data_path.exists());
-    }
-
-    #[test]
-    fn startup_cleanup_removes_an_entire_cache_root() {
-        let directory = tempfile::tempdir().unwrap();
-        let root = directory.path().join("shared");
-        let orphan = root.join("orphaned-viewer");
-        std::fs::create_dir_all(&orphan).unwrap();
-        std::fs::write(orphan.join("attachment.bin"), b"plaintext").unwrap();
-
-        clear_attachment_cache_directory(&root).unwrap();
-
-        assert!(!root.exists());
-        clear_attachment_cache_directory(&root).unwrap();
-    }
-
-    #[test]
-    fn cache_file_guard_removes_abandoned_files_only() {
-        let directory = tempfile::tempdir().unwrap();
-        let abandoned = directory.path().join("abandoned.anb1");
-        std::fs::write(&abandoned, b"partial").unwrap();
-        drop(CacheFileGuard::new(abandoned.clone()));
-        assert!(!abandoned.exists());
-
-        let retained = directory.path().join("retained.anb1");
-        std::fs::write(&retained, b"complete").unwrap();
-        CacheFileGuard::new(retained.clone()).disarm();
-        assert!(retained.exists());
-    }
-
-    #[tokio::test]
-    async fn startup_cleanup_removes_only_preview_cache_scopes() {
-        let directory = tempfile::tempdir().unwrap();
-        let attachment_sync_root = directory.path().join("attachment-sync");
-        let preview_root = attachment_sync_root.join("shared-preview");
-        let durable_root = attachment_sync_root.join("shared");
-        let orphan_scope = preview_root.join("orphan-scope");
-        let durable_scope = durable_root.join("viewer-scope");
-        std::fs::create_dir_all(&orphan_scope).unwrap();
-        std::fs::create_dir_all(&durable_scope).unwrap();
-        std::fs::write(orphan_scope.join("attachment.bin"), b"preview").unwrap();
-        std::fs::write(durable_scope.join("attachment.bin"), b"durable").unwrap();
-
-        assert!(
-            clear_shared_attachment_preview_scopes_at(&preview_root)
-                .await
-                .unwrap()
-        );
-        assert!(!preview_root.exists());
-        assert_eq!(
-            std::fs::read(durable_scope.join("attachment.bin")).unwrap(),
-            b"durable"
-        );
-        assert!(
-            !clear_shared_attachment_preview_scopes_at(&preview_root)
-                .await
-                .unwrap()
-        );
-    }
-}
+mod tests;
