+++ plugins/attachment-sync/src/runtime/tests/mod.rs
@@ -1,0 +1,5 @@
+use super::*;
+
+mod delete_guard;
+mod download;
+mod transfer;
